Create Syllogistic Commonsense Reasoning
Syllogism is a common form of deductive reasoning that requires precisely two premises and one conclusion. It is considered as a logical method to arrive at new information. For instance, given that "Avicenna wrote the famous book the Canon of Medicine " and " The Canon of Medicine has influenced modern medicine," it can be concluded that "Avicenna has influenced modern medicine."
The Avicenna corpus is a benchmark for syllogistic NLI and syllogistic NLG:
Syllogistic NLI: Identifying the possibility of inferring between pairs of inputted sentences.
Syllogistic NLG: Generating a conclusion sentence for two sentences with a syllogistic relation.